+++ NEWS.md
@@ -1,3 +1,9 @@
+## chromConverter 0.6.5
+
+* Added preliminary support for Varian Worktation (`.sms`) format through `read_varian_sms` function.
+* Fixed bug affecting some `mdf` files lacking null bytes after the file header.
+* Sped up `read_shimadzu_lcd` by dealing with twos-complements more sensibly.
+
 ## chromConverter 0.6.4
 
 * Added support for Agilent ChemStation' version 8 (`.ch`) files through `read_chemstation_ch`.